refactor: use export language instead of sync
